Add drop_util file and move function from drop_init
Add function to sync a drop
